Bring new !role in place
